% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.lvim.R
\name{format.lvim}
\alias{format.lvim}
\title{Format a \code{lvim} object}
\usage{
\method{format}{lvim}(x, ...)
}
\arguments{
\item{x}{the \code{lvim} object of interest}

\item{...}{other options, see the generic \code{format} function}
}
\description{
Format a \code{lvim} object
}
