% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvim.R
\name{lvim}
\alias{lvim}
\title{Create a Longitudinal Variable Importance Object}
\usage{
lvim(..., timepoints = numeric())
}
\arguments{
\item{...}{a comma-separated list of individual, cross-sectional variable importance objects. Assumed
to be in order over time.}
}
\value{
an object of class \code{lvim}
}
\description{
Create a longitudinal variable importance object from several constituent
cross-sectional variable importance objects.
}
