% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvim_average.R
\name{lvim_average}
\alias{lvim_average}
\title{Average Longitudinal Variable Importance}
\usage{
lvim_average(lvim, indices = 1:length(lvim), delta = 0)
}
\arguments{
\item{lvim}{an object of class \code{lvim} containing the cross-sectional
variable importance objects}

\item{indices}{a numeric vector indicating the contiguous subset of the
time series}

\item{delta}{null hypothesis value}
}
\value{
The \code{lvim} object, with point estimates, CIs, and p-values
  related to the average variable importance filled in.
}
\description{
Compute a nonparametric estimate of (and efficient influence function for) the
average longitudinal variable importance over a contiguous subset of the time series.
}
