% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvim_trend.R
\name{lvim_trend}
\alias{lvim_trend}
\title{Linear Trend in the Longitudinal Variable Importance Trajectory}
\usage{
lvim_trend(lvim, indices = 1:length(lvim), delta = 0)
}
\arguments{
\item{lvim}{an object of class \code{lvim} containing the cross-sectional
variable importance objects}

\item{indices}{a numeric vector indicating the contiguous subset of the
time series}

\item{delta}{null hypothesis value}
}
\value{
The \code{lvim} object, with point estimates, CIs, and p-values
related to the linear trend in variable importance filled in.
}
\description{
Compute a nonparametric estimate of (and efficient influence function for) the
linear trend in the longitudinal variable importance over a contiguous subset of the time series.
}
