% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvimp-package.R
\docType{package}
\name{lvimp-package}
\alias{lvimp}
\alias{lvimp-package}
\title{lvimp: Perform Inference on Summaries of Longidutinal Algorithm-Agnostic Variable Importance}
\description{
Calculate point estimates of and valid confidence intervals for longitudinal summaries of nonparametric, algorithm-agnostic variable importance measures. For more details, see Williamson et al. (2024) \url{https://arxiv.org/abs/2311.01638}
}
\section{Authors}{

\bold{Maintainer}: Brian Williamson \url{https://bdwilliamson.github.io/}

Methodology authors:
\itemize{
  \item{Brian D. Williamson}
  \item{Susan M. Shortreed}
}
}

\section{See Also}{

Manuscripts: (to appear)
}

\section{Imports}{

The packages that we import either make internal code nice
(dplyr, magrittr, tibble, rlang, data.table), or are used for estimating
and performing inference on cross-sectional variable importance (vimp).

We suggest several other packages: ggplot2 and cowplot help with plotting
variable importance estimates; testthat and covr help with unit tests;
and knitr and rmarkdown help with the vignettes and examples.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://bdwilliamson.github.io/lvimp/}
  \item Report bugs at \url{https://github.com/bdwilliamson/lvimp/issues}
}

}
\author{
\strong{Maintainer}: Brian D. Williamson \email{brian.d.williamson@kp.org} (\href{https://orcid.org/0000-0002-7024-548X}{ORCID})

}
\keyword{internal}
