% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lvim.R
\name{print.lvim}
\alias{print.lvim}
\title{Print a \code{lvim} object}
\usage{
\method{print}{lvim}(x, ...)
}
\arguments{
\item{x}{the \code{lvim} object of interest}

\item{...}{other options, see the generic \code{print} function}
}
\value{
No return value, called for side effects.
}
\description{
Print a \code{lvim} object
}
